refactored the code moving api functions to new file
