+++ src/App.js
@@ -1,39 +1,19 @@
 import React, { useEffect, useState } from 'react';
 import TodoList from './TodoList';
 import AddTodoForm from './AddTodoForm';
+import { fetchTodosFromAPI, addTodoToAPI } from './apiFunctions'
 
 function App() {
   const [todoList, setTodoList] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
 
-  async function fetchData() {
-    const options = {
-      method: 'GET',
-      headers: { 'Authorization': `Bearer ${process.env.REACT_APP_AIRTABLE_API_TOKEN}` }
-    };
-    const url = `https://api.airtable.com/v0/${process.env.REACT_APP_AIRTABLE_BASE_ID}/${process.env.REACT_APP_TABLE_NAME}`;
-
-    try {
-      const response = await fetch(url, options);
-      if (!response.ok) {
-        const message = `Error: ${response.status}`;
-        throw new Error(message);
-      }
-      const data = await response.json()
-      console.log(data)
-      const todos = data.records.map((todo) => ({
-        id: todo.id,
-        title: todo.fields.title
-      }));
-      console.log(todos)
-      setTodoList(todos);
-      setIsLoading(false);
-    } catch (err) {
-      console.log(err.message)
-    }
-  }
-
   useEffect(() => {
+    async function fetchData() {
+      const todos = await fetchTodosFromAPI();
+      console.log(todos)
+      setTodoList([...todos]);
+      setIsLoading(false);
+    }
     fetchData();
   }, []);
 
@@ -44,32 +24,15 @@ function App() {
   }, [todoList, isLoading]);
 
   async function addTodo(newTodo) {
-    try {
-      const options = {
-        method: 'POST',
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${process.env.REACT_APP_AIRTABLE_API_TOKEN}`,
-        },
-        body: JSON.stringify({ fields: { title: newTodo.title } }),
-      };
-      const url = `https://api.airtable.com/v0/${process.env.REACT_APP_AIRTABLE_BASE_ID}/${process.env.REACT_APP_TABLE_NAME}`;
-
-      const response = await fetch(url, options);
-      if (!response.ok) {
-        throw new Error(`Error:${response.status}`)
-      }
-
-      const addTodo = await response.json()
-      setTodoList([...todoList, { id: addTodo.id, title: addTodo.fields.title }]);
-
-    } catch (err) {
-      console.log(err.message)
+    const addedTodo = await addTodoToAPI(newTodo);
+    if (addedTodo) {
+      setTodoList([...todoList, addedTodo])
     }
+
     //setTodoList([...todoList, newTodo])
   }
 
-  function removeTodo(id) {
+  function removeTodoFromList(id) {
     const updatedTodoList = todoList.filter(todo => todo.id !== id)
     setTodoList(updatedTodoList);
   }
@@ -82,7 +45,7 @@ function App() {
           <p>Loading...</p>
         ) : (
           <>
-            <TodoList todoList={todoList} onRemoveTodo={removeTodo} />
+            <TodoList todoList={todoList} onRemoveTodo={removeTodoFromList} />
           </>
         )}
         <AddTodoForm onAddTodo={addTodo} />
